requirements.txt
numpy
pandas
pyyaml
pydantic
scikit-learn
matplotlib